requirements.txt
jupiter
chardet
pyyaml
